+++ shuup/core/__init__.py
@@ -6,6 +6,7 @@
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
 from shuup.apps import AppConfig
+from shuup.core.excs import MissingSettingException
 
 
 class ShuupCoreAppConfig(AppConfig):
@@ -26,6 +27,13 @@ class ShuupCoreAppConfig(AppConfig):
             "shuup.core.pricing.default_pricing:DefaultPricingModule"
         ],
     }
+
+    def ready(self):
+        from django.conf import settings
+        if not getattr(settings, "PARLER_DEFAULT_LANGUAGE_CODE", None):
+            raise MissingSettingException("PARLER_DEFAULT_LANGUAGE_CODE must be set.")
+        if not getattr(settings, "PARLER_LANGUAGES", None):
+            raise MissingSettingException("PARLER_LANGUAGES must be set.")
 
 
 default_app_config = "shuup.core.ShuupCoreAppConfig"